+++ src/mvc/decorator/Service.ts
@@ -1,3 +1,0 @@
-import {Component} from "../../di/decorator/Component";
-
-export const Service = Component;