+++ changelog.md
@@ -5,4 +5,5 @@
 * add `files` to ApplicationSettings, remove `srcDir, publicDir, logDir, configDir, dbDir`, `rootDir` is not required
 * use `SettingOptions` in `ApplicationLoader` constructor
 * use `init()` return server inside LOON
+* remove `@Service()` decorator